+++ src/main/java/com/baichen/rpc/message/HeartbeatRequest.java
@@ -1,0 +1,8 @@
+package com.baichen.rpc.message;
+
+import lombok.Data;
+
+@Data
+public class HeartbeatRequest {
+    private final long requestTime = System.currentTimeMillis();
+}