+++ src/main/java/com/baichen/rpc/message/HeartbeatResponse.java
@@ -1,0 +1,12 @@
+package com.baichen.rpc.message;
+
+import lombok.Data;
+
+@Data
+public class HeartbeatResponse {
+    private final long requestTime;
+
+    public HeartbeatResponse(long requestTime) {
+        this.requestTime = requestTime;
+    }
+}